pyproject.toml
[project]
name = "advanced-hybrid-rag"
version = "0.1.0"
description = "Hybrid RAG pipeline with multi-modal document processing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LLM & Embeddings
    "ollama>=0.4.0",
    "llama-index-core>=0.12.0",
    "llama-index-llms-ollama>=0.4.0",
    "llama-index-embeddings-ollama>=0.4.0",
    "llama-index-vector-stores-qdrant>=0.4.0",
    "llama-index-postprocessor-cohere-rerank>=0.3.0",

    # Vector Store
    "qdrant-client>=1.12.0",

    # Text Processing & Chunking
    "langchain-experimental>=0.3.0",
    "langchain-community>=0.3.0",
    "tiktoken>=0.8.0",
    "markdown>=3.7",
    "beautifulsoup4>=4.12.0",
    "json-repair>=0.30.0",

    # Document Processing
    "pdf2image>=1.17.0",
    "docling-core>=2.0.0",
    "mlx-vlm>=0.1.0",

    # Computer Vision
    "ultralytics>=8.3.0",
    "opencv-python>=4.10.0",
    "pillow>=11.0.0",

    # Azure OpenAI (optional)
    "openai>=1.50.0",
    "tenacity>=9.0.0",

    # Web UI
    "gradio>=5.0.0",

    # Utilities
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "torch>=2.5.0",
    "ipython>=8.29.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "pylint>=3.3.0",
]
